+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v3.4.3 (unreleased)
  - Add optional 'columns' tag to Index definitions
+ - update dosa-idl into v3.2.1
+ - add columns support to materialized view creation using Schema dump
 
 ## v3.4.2 (2019-03-26)
  - Update version.VERSION